backend/requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
Pillow==6.2.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
Markdown==3.3.3
networkx==2.5
nose==1.3.7
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-bidi==0.4.2
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
